Rename docs, examples to doc, example.
These directories are collections. They should use singular form.
That is more consistent with the style the Go project uses.

Fixes #128.
