+++ example/todomvc/components/filterbutton.go
@@ -4,10 +4,10 @@ import (
 	"github.com/gopherjs/vecty"
 	"github.com/gopherjs/vecty/elem"
 	"github.com/gopherjs/vecty/event"
-	"github.com/gopherjs/vecty/examples/todomvc/actions"
-	"github.com/gopherjs/vecty/examples/todomvc/dispatcher"
-	"github.com/gopherjs/vecty/examples/todomvc/store"
-	"github.com/gopherjs/vecty/examples/todomvc/store/model"
+	"github.com/gopherjs/vecty/example/todomvc/actions"
+	"github.com/gopherjs/vecty/example/todomvc/dispatcher"
+	"github.com/gopherjs/vecty/example/todomvc/store"
+	"github.com/gopherjs/vecty/example/todomvc/store/model"
 	"github.com/gopherjs/vecty/prop"
 )
