+++ example/todomvc/example.go
@@ -6,11 +6,11 @@ import (
 
 	"github.com/gopherjs/gopherjs/js"
 	"github.com/gopherjs/vecty"
-	"github.com/gopherjs/vecty/examples/todomvc/actions"
-	"github.com/gopherjs/vecty/examples/todomvc/components"
-	"github.com/gopherjs/vecty/examples/todomvc/dispatcher"
-	"github.com/gopherjs/vecty/examples/todomvc/store"
-	"github.com/gopherjs/vecty/examples/todomvc/store/model"
+	"github.com/gopherjs/vecty/example/todomvc/actions"
+	"github.com/gopherjs/vecty/example/todomvc/components"
+	"github.com/gopherjs/vecty/example/todomvc/dispatcher"
+	"github.com/gopherjs/vecty/example/todomvc/store"
+	"github.com/gopherjs/vecty/example/todomvc/store/model"
 )
 
 func main() {
