+++ example/todomvc/store/model/model.go
